2.0 menus

+++ SkyblockCore-Common/src/main/java/me/illusion/skyblockcore/common/utilities/time/Time.java
@@ -21,8 +21,8 @@ public class Time {
      * @param unit The unit to convert to
      * @return The converted time
      */
-    public long as(TimeUnit unit) {
-        return unit.convert(duration, this.unit);
+    public int as(TimeUnit unit) {
+        return (int) unit.convert(duration, this.unit);
     }
 
     /**
